+++ dhareshsawLescargot/src/app/app-routing.module.ts
@@ -18,11 +18,7 @@ const routes: Routes = [
   {
     path: 'object-inventory',
     loadChildren: () => import('./object-inventory/object-inventory.module').then( m => m.ObjectInventoryPageModule)
-  },  {
-    path: 'home-page',
-    loadChildren: () => import('./home-page/home-page.module').then( m => m.HomePagePageModule)
   },
-
 ];
 
 @NgModule({
