+++ dhareshsawLescargot/src/app/classes/objectInventory.ts
@@ -1,0 +1,8 @@
+export interface ObjectInventory {
+
+    // un objet est composé d'un nom et d'un pouvoir ==> vu ac johann le 13MAI 
+    nameObject: string,
+    power: string,
+    image: string,
+    
+  }
