+++ dhareshsawLescargot/src/app/datas/listePersonnages.ts
@@ -1,0 +1,22 @@
+import { Character } from '../classes/personnage';
+
+export const PERSONNAGES: Character[] = [
+  {
+    _id: '0',
+    name: 'Dhareshsaw',
+    avatar: './assets/avatars/dhareshsaw.png',
+    strength: 4,
+    luck: 3,
+    endurance: 2,
+    items: null,
+  },
+  {
+    _id: '1',
+    name: 'Jojo le bourricot',
+    avatar: './assets/avatars/johann.png',
+    strength: 3,
+    luck: 2,
+    endurance: 4,
+    items: null,
+  }
+]
