--- dhareshsawLescargot/src/app/listePersonnages.ts
@@ -1,22 +1,0 @@
-import { Character } from './personnage';
-
-export const PERSONNAGES: Character[] = [
-  {
-    _id: '0',
-    name: 'Dhareshsaw',
-    avatar: './assets/avatars/dhareshsaw.png',
-    strength: 4,
-    luck: 3,
-    endurance: 2,
-    items: null,
-  },
-  {
-    _id: '1',
-    name: 'Jojo le bourricot',
-    avatar: './assets/avatars/johann.png',
-    strength: 3,
-    luck: 2,
-    endurance: 4,
-    items: null,
-  }
-]
