+++ dhareshsawLescargot/src/app/services/scene.service.ts
@@ -1,4 +1,5 @@
 import { Injectable } from '@angular/core';
+import { SCENES } from '../datas/listeScenes';
 
 @Injectable({
   providedIn: 'root'
@@ -6,4 +7,9 @@ import { Injectable } from '@angular/core';
 export class SceneService {
 
   constructor() { }
+
+  getSceneById(id){
+    return SCENES.find(({_id})=> _id === id);
+  }
+
 }
